cleanup: move interpolate inline function to its own header file
This is not really related to struct dive, so let's move it.
